replacing vhost add/del old versions

+++ bin/menu/vhost-del.sh
@@ -2,7 +2,7 @@
 
 if [ $# -lt 1 ]; then
 	echo "----"
-	echo "$0 | Deletes 'domain.com' virtual host to apache."
+	echo "$0 | Deletes 'domain.com' virtual host to apache (ubuntu 14)."
 	echo "Usage: $0 domain.com"
 	echo "----"
 	exit 1
@@ -11,8 +11,15 @@ fi
 
 domain="$1"
 DOCROOT="/var/www/vhosts/${domain}"
+VER=$(grep RELEASE /etc/lsb-release |cut -d\= -f2|cut -d\. -f1)
 
-   	 rm  /etc/apache2/sites-available/"${domain}"
+        if [ $VER -gt 13 ]
+                then
+	                 rm  /etc/apache2/sites-available/"${domain}.conf"
+                 else
+                 	rm  /etc/apache2/sites-available/"${domain}"
+                 fi
+              
          rmdir ${DOCROOT}
          echo "Disabling virtual host..."	 
          a2dissite $domain > /dev/null 2>&1
